be-ucar/requirements.txt
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
databases==0.7.0
fastapi==0.89.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
